requirements.txt
aiogram==3.*
python-dotenv==1.*
SQLAlchemy==2.*
aiosqlite==0.20.*